fix(nx-oc): honor --no-interactive for the sandbox registry prompt
The sandbox generator's registry fallback (--registry → nx.json → derived from
git remote → enquirer prompt) opened the prompt even under --no-interactive:
Nx's flag only suppresses its own schema x-prompts, not manual enquirer prompts.
So a non-interactive/CI run with no --registry and no derivable git remote hung
on "What container registry should sandbox images be published to?".

Guard the prompt with isNonInteractive() — throw an actionable error naming
--registry=<host>/<org> instead of blocking. Extracts isNonInteractive() into
utils/interactive.ts (shared with adsp-utils, which had a private copy).

Regression test: no --registry + no git remote + non-interactive (jest has no
TTY) now rejects with a /--registry/ error instead of hanging. nx-oc build/
test (101)/lint green.
